feat: Change history table to toplines
