+++ src/components/pages/state/long-term-care/chart.js
@@ -5,6 +5,7 @@ import LineChart from '~components/charts/line-chart'
 import colors from '~scss/colors.module.scss'
 import chartStyles from './chart.module.scss'
 import Alert from '~components/common/alert'
+import getTotals from '~utilities/ltc-totals'
 
 const chartProps = {
   height: 180, // these control the dimensions used to render the svg but not the final size
@@ -14,64 +15,6 @@ const chartProps = {
   marginRight: 30,
   xTicks: 3,
   showTicks: 6,
-}
-
-const getTotals = state => {
-  const categories = ['nh', 'alf', 'lumpedother']
-  let totalCases = null
-  let totalDeath = null
-  categories.forEach(category => {
-    if (
-      state[`posresstaff_${category}`] ||
-      state[`probposresstaff_${category}`]
-    ) {
-      if (state[`posresstaff_${category}`]) {
-        totalCases += state[`posresstaff_${category}`]
-      }
-      if (state[`probposresstaff_${category}`]) {
-        totalCases += state[`probposresstaff_${category}`]
-      }
-    } else {
-      if (state[`posres_${category}`]) {
-        totalCases += state[`posres_${category}`]
-      }
-      if (state[`probposres_${category}`]) {
-        totalCases += state[`probposres_${category}`]
-      }
-      if (state[`posstaff_${category}`]) {
-        totalCases += state[`posstaff_${category}`]
-      }
-      if (state[`probposstaff_${category}`]) {
-        totalCases += state[`probposstaff_${category}`]
-      }
-    }
-    if (
-      state[`deathresstaff_${category}`] ||
-      state[`probdeathresstaff_${category}`]
-    ) {
-      if (state[`deathresstaff_${category}`]) {
-        totalDeath += state[`deathresstaff_${category}`]
-      }
-      if (state[`probdeathresstaff_${category}`]) {
-        totalDeath += state[`probdeathresstaff_${category}`]
-      }
-    } else {
-      if (state[`deathres_${category}`]) {
-        totalDeath += state[`deathres_${category}`]
-      }
-      if (state[`probdeathres_${category}`]) {
-        totalDeath += state[`probdeathres_${category}`]
-      }
-      if (state[`deathstaff_${category}`]) {
-        totalDeath += state[`deathstaff_${category}`]
-      }
-      if (state[`probdeathstaff_${category}`]) {
-        totalDeath += state[`probdeathstaff_${category}`]
-      }
-    }
-  })
-
-  return { totalCases, totalDeath }
 }
 
 const LongTermCareCharts = ({ data }) => {
