+++ src/components/pages/state/long-term-care/overview.js
@@ -3,11 +3,11 @@ import { useStaticQuery, graphql } from 'gatsby'
 import { Row, Col } from '~components/common/grid'
 import Total from '~components/common/landing-page/total'
 import { DefinitionPanel } from '~components/pages/data/cards/definitions-panel'
-import { CtaLink, CtaAnchorLink } from '~components/common/call-to-action'
+import { CtaAnchorLink } from '~components/common/call-to-action'
 import { FormatNumber } from '~components/utils/format'
 import overviewStyles from './overview.module.scss'
 
-const LongTermCareOverview = ({ facilities, overview, stateSlug }) => {
+const LongTermCareOverview = ({ facilities, overview }) => {
   const [highlightedDefinition, setHighlightedDefinition] = useState(false)
   const { allContentfulDataDefinition } = useStaticQuery(graphql`
     {
@@ -101,9 +101,6 @@ const LongTermCareOverview = ({ facilities, overview, stateSlug }) => {
           <CtaAnchorLink block href="#facilities">
             All facilities
           </CtaAnchorLink>
-          <CtaLink to={`/data/state/${stateSlug}/long-term-care/history`} block>
-            Historical totals
-          </CtaLink>
         </Col>
       </Row>
     </>
